crawler-main-tzb/requirements.txt
selenium==3.141.0
psutil==5.8.0
requests==2.25.1
beautifulsoup4==4.9.3
Pillow==8.2.0
lxml==4.6.3
python-dateutil==2.8.2
